added new scenarios
