+++ pom.xml
@@ -104,8 +104,15 @@
             <version>4.12</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.hamcrest</groupId>
+            <artifactId>hamcrest-all</artifactId>
+            <version>1.3</version>
+            <scope>test</scope>
+        </dependency>
 
-                <!--Serenity Dependencies-->
+
+        <!--Serenity Dependencies-->
 
         <!--Groovy Dependencies-->
         <dependency>
