--- src/test/java/SignUpJava.java
@@ -1,35 +1,0 @@
-import io.restassured.http.ContentType;
-import net.serenitybdd.rest.SerenityRest;
-import net.serenitybdd.screenplay.Actor;
-import net.serenitybdd.screenplay.Performable;
-import net.serenitybdd.screenplay.Question;
-import net.serenitybdd.screenplay.Task;
-import net.serenitybdd.screenplay.questions.converters.IntegerConverter;
-import net.serenitybdd.screenplay.rest.interactions.Get;
-import net.serenitybdd.screenplay.rest.interactions.Post;
-
-import static net.serenitybdd.screenplay.Tasks.instrumented;
-
-public class SignUpJava implements Task {
-
-    private final Object userInfo;
-
-    public SignUpJava(Object userInfo) {
-        this.userInfo = userInfo;
-    }
-
-    public static Performable withInfo(Object userInfo) {
-        return instrumented(SignUpJava.class, userInfo);
-    }
-
-    @Override
-    public <T extends Actor> void performAs(T actor) {
-      /*  actor.attemptsTo(
-                Post.to("/api/users").with(
-                        requestSpecification -> requestSpecification
-                                .contentType(ContentType.JSON)
-                                .body(userInfo)
-                )
-        );*/
-    }
-}
